earnings-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pymongo>=4.6.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiofiles>=23.2.0
